+++ src/Rental.Core/Models/Rentals/RentalWithPaymentDetails.cs
@@ -8,9 +8,9 @@ namespace Rental.Core.Models.Rentals
         public Guid Id { get; set; }
         public string ClientName { get; set; }
         public string BoardGameName { get; set; }
-        public float ChargedDeposit { get; set; }
+        public decimal ChargedDeposit { get; set; }
         public DateTime RentalStartDateTime { get; set; }
-        public float MoneyToPay { get; set; }
+        public decimal MoneyToPay { get; set; }
         public ICollection<RentalDay> RentalDays { get; set; }
     }
 }
